+++ settings.gradle
@@ -37,3 +37,4 @@ include 'integration-test'
 include 'supplier'
 include 'supplier:local'
 include 'supplier:ogmios-supplier'
+include 'supplier:kupo-supplier'
