+++ supplier/build.gradle
@@ -1,6 +1,7 @@
 dependencies {
     api project(':supplier:local')
     api project(':supplier:ogmios-supplier')
+    api project(':supplier:kupo-supplier')
 }
 
 publishing {
