+++ supplier/kupo-supplier/README.md
@@ -1,0 +1,3 @@
+## Kupo Supplier Module (cardano-client-supplier-ogmios)
+
+Kupo supplier to deliver utxo functionalities via Kupo API.
